+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.0.1
+
+- Remove advancements.
+- Add charmony icon.
+- Update to use the new registry methods.
+- Bees no longer try to pollinate baby mooblooms.
+
 ## 1.0.0
 
-Initial version.
+- Initial version.